IC703 controls
  * Added controls and supporting methods for
    - AF level
    - RF level
    - Squelch
    - IF shift
    - RF power
    - Mic gain
    - CW pitch
    - CW WPM
    - CW sidetone
  * Added read support for ALC
  * Added Compression level control
  * Added VOX gain, anti and delay controls

+++ src/include/status.h
@@ -92,7 +92,7 @@ struct status {
 	int		cw_spot;
 	bool	spot_onoff;
 	int		cw_spot_tone;
-	int		cw_qsk;
+	double	cw_qsk;
 	bool	enable_keyer;
 	bool	vox_onoff;
 	int		vox_gain;
